Safari-609.1.11.1/LayoutTests/http/tests/security/window-name-after-cross-origin-main-frame-navigation.html

<!DOCTYPE html>
<html>
    <script>
        if (window.testRunner) {
            testRunner.dumpAsText();
            testRunner.waitUntilDone();
        }

        console.log("window.name should equal \"\" after a cross-origin main frame navigation.");
        document.location = "resources/window-name-test.html?localhost";
    </script>
</html>
