Safari-609.1.11.1/LayoutTests/http/tests/security/window-name-after-cross-origin-main-frame-navigation-expected.txt

CONSOLE MESSAGE: line 9: window.name should equal "" after a cross-origin main frame navigation.
CONSOLE MESSAGE: line 4: window.name equals "".
